requirements.txt
boto3>=1.34,<2
PyYAML>=6.0,<7
